fix(plugin-sdk): preserve cancellation receivers
