+++ extensions/discord/src/monitor/message-handler.ingress-recovery.test.ts
@@ -1,7 +1,4 @@
 // Discord tests cover durable retry recovery through full handler replacement.
-import fs from "node:fs/promises";
-import os from "node:os";
-import path from "node:path";
 import type { APIMessage } from "discord-api-types/v10";
 import {
   type ChannelIngressQueue,
@@ -12,7 +9,8 @@ import {
   closeOpenClawStateDatabaseForTest,
   createChannelIngressQueueForTests,
 } from "openclaw/plugin-sdk/plugin-state-test-runtime";
-import { describe, expect, it, vi } from "vitest";
+import { afterEach, describe, expect, it, vi } from "vitest";
+import { useAutoCleanupTempDirTracker } from "../../../../test/helpers/temp-dir.js";
 import { createDiscordIngressMonitor } from "./ingress.js";
 import { createDiscordMessageHandler } from "./message-handler.js";
 import { createDiscordHandlerParams } from "./message-handler.test-helpers.js";
@@ -24,6 +22,8 @@ type DiscordIngressPayload = {
 };
 
 type DiscordQueue = ChannelIngressQueue<DiscordIngressPayload>;
+
+const tempDirs = useAutoCleanupTempDirTracker(afterEach);
 
 function rawMessage(id: string, channelId = "lane-a"): APIMessage {
   return {
@@ -51,8 +51,7 @@ function rawMessage(id: string, channelId = "lane-a"): APIMessage {
 }
 
 async function withQueue(run: (queue: DiscordQueue) => Promise<void>): Promise<void> {
-  const created = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-discord-recovery-"));
-  const stateDir = await fs.realpath(created);
+  const stateDir = tempDirs.make("openclaw-discord-recovery-");
   const queue = createChannelIngressQueueForTests<DiscordIngressPayload>({
     channelId: "discord",
     accountId: "default",
@@ -62,7 +61,6 @@ async function withQueue(run: (queue: DiscordQueue) => Promise<void>): Promise<v
     await run(queue);
   } finally {
     closeOpenClawStateDatabaseForTest();
-    await fs.rm(stateDir, { recursive: true, force: true });
   }
 }
