+++ extensions/discord/src/monitor/message-handler.queue.test.ts
@@ -1,8 +1,5 @@
 // Discord tests cover message handler.queue plugin behavior.
 import { getEventListeners } from "node:events";
-import fs from "node:fs/promises";
-import os from "node:os";
-import path from "node:path";
 import type { APIMessage } from "discord-api-types/v10";
 import {
   type ChannelIngressQueue,
@@ -14,7 +11,8 @@ import {
   closeOpenClawStateDatabaseForTest,
   createChannelIngressQueueForTests,
 } from "openclaw/plugin-sdk/plugin-state-test-runtime";
-import { beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { useAutoCleanupTempDirTracker } from "../../../../test/helpers/temp-dir.js";
 import { createDiscordIngressMonitor, type DiscordIngressLifecycle } from "./ingress.js";
 import { createDiscordMessageHandler as createDurableDiscordMessageHandler } from "./message-handler.js";
 import {
@@ -68,11 +66,12 @@ type DiscordIngressPayload = {
   rawMessage: APIMessage;
 };
 
+const tempDirs = useAutoCleanupTempDirTracker(afterEach);
+
 async function withDiscordQueue<T>(
   run: (queue: ChannelIngressQueue<DiscordIngressPayload>) => Promise<T>,
 ): Promise<T> {
-  const created = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-discord-handler-"));
-  const stateDir = await fs.realpath(created);
+  const stateDir = tempDirs.make("openclaw-discord-handler-");
   const queue = createChannelIngressQueueForTests<DiscordIngressPayload>({
     channelId: "discord",
     accountId: "default",
@@ -82,7 +81,6 @@ async function withDiscordQueue<T>(
     return await run(queue);
   } finally {
     closeOpenClawStateDatabaseForTest();
-    await fs.rm(stateDir, { recursive: true, force: true });
   }
 }
