+++ src/plugin-sdk/channel-ingress-runtime.ts
@@ -153,17 +153,16 @@ export function fanInChannelIngressLifecycles(
   const failAll = async (error: unknown) => {
     await Promise.all(lifecycles.map(async (lifecycle) => await lifecycle.onFailed?.(error)));
   };
-  const cancellationHandlers = lifecycles.flatMap((lifecycle) =>
-    lifecycle.onCancelled ? [lifecycle.onCancelled] : [],
-  );
+  const supportsCancellation = lifecycles.every((lifecycle) => lifecycle.onCancelled !== undefined);
   // Omit aggregate cancellation unless every durable source supports it. Callers
   // can then use settle/abandon without an acknowledged-but-unsettled claim.
-  const cancelAll =
-    cancellationHandlers.length === lifecycles.length
-      ? async () => {
-          await Promise.all(cancellationHandlers.map(async (cancel) => await cancel()));
-        }
-      : undefined;
+  const cancelAll = async () => {
+    await Promise.all(
+      lifecycles.map(async (lifecycle) =>
+        lifecycle.onCancelled ? await lifecycle.onCancelled() : await lifecycle.onAbandoned(),
+      ),
+    );
+  };
   return {
     lifecycle: {
       abortSignal:
@@ -189,7 +188,7 @@ export function fanInChannelIngressLifecycles(
         handedOff = true;
         await failAll(error);
       },
-      ...(cancelAll
+      ...(supportsCancellation
         ? {
             onCancelled: async () => {
               handedOff = true;
@@ -220,11 +219,7 @@ export function fanInChannelIngressLifecycles(
     cancel: async () => {
       if (!handedOff) {
         handedOff = true;
-        await Promise.all(
-          lifecycles.map(async (lifecycle) =>
-            lifecycle.onCancelled ? await lifecycle.onCancelled() : await lifecycle.onAbandoned(),
-          ),
-        );
+        await cancelAll();
       }
     },
   };
